feat(sensor): merge test files for config and env
